Prevent fold from being optimized away entirely.
Remove the semi-colon on the fold so it's result is returned to the
bencher. Without this all of the fold functions were being removed by
the optimizer as it determined that their results were unused.

Now that the folds are actually getting generated I get an LLVM
error:

```
LLVM ERROR: Cannot select: intrinsic %llvm.x86.sse41.dpps
```